refactor: move Spark-coupled adapter code into a spark subpackage
