+++ src/delta_engine/adapters/databricks/spark/__init__.py
@@ -1,0 +1,9 @@
+"""
+Spark backend for the Databricks adapter.
+
+Everything PySpark-coupled lives here: the reader, the executor, the engine
+factory, and py4j error translation. Import concrete modules directly
+(``factory``, ``reader``, ``executor``, ``errors``); the public user-facing
+entry points live in :mod:`delta_engine.databricks`. This ``__init__`` stays
+empty so importing the package never pulls PySpark.
+"""
