+++ src/delta_engine/adapters/databricks/spark/errors.py
