+++ src/delta_engine/adapters/databricks/spark/executor.py
@@ -10,7 +10,7 @@ import logging
 
 from pyspark.sql import SparkSession
 
-from delta_engine.adapters.databricks.errors import summarize_exception
+from delta_engine.adapters.databricks.spark.errors import summarize_exception
 from delta_engine.adapters.databricks.sql import compile_plan
 from delta_engine.application.failures import ExecutionFailure
 from delta_engine.application.ports import (
