+++ src/delta_engine/adapters/databricks/spark/reader.py
@@ -12,7 +12,7 @@ from pyspark.sql import Row, SparkSession
 from pyspark.sql.catalog import Column as SparkColumn
 from pyspark.sql.types import DataType as SparkType
 
-from delta_engine.adapters.databricks.errors import summarize_exception
+from delta_engine.adapters.databricks.spark.errors import summarize_exception
 from delta_engine.adapters.databricks.sql import (
     column_tags_query,
     describe_detail_query,
