+++ tests/adapters/databricks/spark/test_build_engine.py
@@ -1,7 +1,7 @@
 import io
 import logging
 
-from delta_engine.adapters.databricks.factory import build_engine
+from delta_engine.adapters.databricks.spark.factory import build_engine
 from delta_engine.adapters.databricks.log_config import LevelColorFormatter, configure_logging
 from delta_engine.application.engine import Engine
