+++ tests/e2e/test_engine_e2e.py
@@ -4,8 +4,8 @@ from uuid import uuid4
 import pyspark.sql.types as T
 import pytest
 
-from delta_engine.adapters.databricks.executor import DatabricksExecutor
-from delta_engine.adapters.databricks.reader import DatabricksReader
+from delta_engine.adapters.databricks.spark.executor import DatabricksExecutor
+from delta_engine.adapters.databricks.spark.reader import DatabricksReader
 from delta_engine.application.engine import Engine
 from delta_engine.application.errors import SyncFailedError
 from delta_engine.application.failures import ValidationFailure
